1) Detect garbage words, 2) insert spaces into concatenated words
Modules for processing words. To be used for Playful Words.

For instance, the garbage detector may be used to pass garbage words
to a second synthesizer with different pronunciation rules.

The string splitter may be used to help detect whether the string was
created with intent, i.e. the child was making words versus just randomly
stringing together letters.

Two main modules:
1) garbage_detector: Given a string, e.g. 'haltlaper', return whether
   it is a garbage word or not. A *nonsense* word is a non-English string
   that looks like an English word, e.g. 'plugort'. A *garbage* word is a
   string that doesn't look like a word at all, e.g. 'mwlrpokz'.

2) string_splitter: Given a string, e.g. 'buzzkgpotterykbarn', insert spaces
   to form 'buzz kg pottery k barn'

2 other files:
- analyze.py: basic analysis on the spoken words.
- process_logs.py: I wrote a small script on the Playful Words computer
  to pickle all the logs. I copied that over. This simply goes through
  the pickled list and writes it to csv files.
    - spoken_words is the unique set of spoken_words_all