[GUI] Ability to click and analyze speechblocks words
List of spoken words are now clickable. Once clicked, the word is
analyzed, and both the current set of analyzed words and the
spoken words list are updated.

Currently, the list of analyzed words isn't automatically updated. I haven't
done this yet because these components don't have a child-parent relationship;
instead of updating through a props callback, I'll have to setup a global event
system.

It does get updated every 2 seconds though because of the polling on line 74.

Notes:
- Helpful link: https://facebook.github.io/react/tips/communicate-between-components.html
- The onclick functionality is done by binding a function that takes the
  index of each candidate.
- To update the list, we have to call setState. However, CandidateList
  doesn't actually have state -- it's CandidateBox that does. Thus,
  CandidateBox simply passes an updateState function that can be accessed
  through the props.
- Aside: There's a weird scope issue where 'this' becomes undefined if
  accessed within the map function on line 153. That's why I create
  a variable called dis.
- To update global variables in Flask routing functions, declare global
  (see line 76)